Rewrite ToSlice return empty slice on no valids
Previously ToSlice would use make to create
a list of same length as input string s.
But whenever an invalid char was encountered,
an "empty" space would be left in the slice.
Get around this by not pre-allocating the slice
and using append. Probably a more elegant
way exists to handle this situation.